Move core logic somewhere else
